Iteration 2
- Essential
Remove all unnecessary literals.

In JavaScript it's not necessary to return the literal true or false in order to return a boolean. The result of an expression can be returned directly.

- Recommended
Simplify or remove if statements.

This exercise is meant to explore booleans and boolean operators. No if statements are required to complete this exercise.